# Backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/meals/": {
      "get": {
        "tags": [
          "meals"
        ],
        "summary": "Get All Meals",
        "description": "Return all meals.",
        "operationId": "get_all_meals_api_meals__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MealRead"
                  },
                  "type": "array",
                  "title": "Response Get All Meals Api Meals  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "meals"
        ],
        "summary": "Add Meal",
        "description": "Create a new meal.",
        "operationId": "add_meal_api_meals__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MealCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/meals/possible_tags": {
      "get": {
        "tags": [
          "meals"
        ],
        "summary": "Get Possible Meal Tags",
        "description": "Return all possible meal tags ordered by name.",
        "operationId": "get_possible_meal_tags_api_meals_possible_tags_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PossibleMealTag"
                  },
                  "type": "array",
                  "title": "Response Get Possible Meal Tags Api Meals Possible Tags Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "meals"
        ],
        "summary": "Create Possible Meal Tag",
        "description": "Create a new possible meal tag.",
        "operationId": "create_possible_meal_tag_api_meals_possible_tags_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PossibleMealTagCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PossibleMealTag"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/meals/possible_tags/{tag_id}": {
      "put": {
        "tags": [
          "meals"
        ],
        "summary": "Update Possible Meal Tag",
        "description": "Update an existing possible meal tag.",
        "operationId": "update_possible_meal_tag_api_meals_possible_tags__tag_id__put",
        "parameters": [
          {
            "name": "tag_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tag Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PossibleMealTagUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PossibleMealTag"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "meals"
        ],
        "summary": "Delete Possible Meal Tag",
        "description": "Delete a possible meal tag if not linked.",
        "operationId": "delete_possible_meal_tag_api_meals_possible_tags__tag_id__delete",
        "parameters": [
          {
            "name": "tag_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tag Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Possible Meal Tag Api Meals Possible Tags  Tag Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/meals/{meal_id}": {
      "get": {
        "tags": [
          "meals"
        ],
        "summary": "Get Meal",
        "description": "Retrieve a single meal by ID.",
        "operationId": "get_meal_api_meals__meal_id__get",
        "parameters": [
          {
            "name": "meal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Meal Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "meals"
        ],
        "summary": "Update Meal",
        "description": "Update an existing meal.",
        "operationId": "update_meal_api_meals__meal_id__put",
        "parameters": [
          {
            "name": "meal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Meal Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MealUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "meals"
        ],
        "summary": "Delete Meal",
        "description": "Delete a meal.",
        "operationId": "delete_meal_api_meals__meal_id__delete",
        "parameters": [
          {
            "name": "meal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Meal Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Meal Api Meals  Meal Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MealCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "ingredients": {
            "items": {
              "$ref": "#/components/schemas/MealIngredientCreate"
            },
            "type": "array",
            "title": "Ingredients"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/TagRef"
            },
            "type": "array",
            "title": "Tags"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "MealCreate",
        "description": "Schema for creating a meal."
      },
      "MealIngredient": {
        "properties": {
          "ingredient_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ingredient Id"
          },
          "meal_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meal Id"
          },
          "unit_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Id"
          },
          "unit_quantity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Quantity"
          }
        },
        "type": "object",
        "title": "MealIngredient",
        "description": "Link between a meal and an ingredient with quantity information."
      },
      "MealIngredientCreate": {
        "properties": {
          "ingredient_id": {
            "type": "integer",
            "title": "Ingredient Id"
          },
          "unit_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Id"
          },
          "unit_quantity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Quantity"
          }
        },
        "type": "object",
        "required": [
          "ingredient_id"
        ],
        "title": "MealIngredientCreate",
        "description": "Schema for creating meal ingredient linkage."
      },
      "MealRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "ingredients": {
            "items": {
              "$ref": "#/components/schemas/MealIngredient"
            },
            "type": "array",
            "title": "Ingredients"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/PossibleMealTag"
            },
            "type": "array",
            "title": "Tags"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "MealRead",
        "description": "Schema for reading meal data."
      },
      "MealUpdate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "ingredients": {
            "items": {
              "$ref": "#/components/schemas/MealIngredientCreate"
            },
            "type": "array",
            "title": "Ingredients"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/TagRef"
            },
            "type": "array",
            "title": "Tags"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "MealUpdate",
        "description": "Schema for updating a meal."
      },
...
      "PossibleMealTag": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "group": {
            "type": "string",
            "title": "Group"
          }
        },
        "type": "object",
        "required": [
          "name",
          "group"
        ],
        "title": "PossibleMealTag",
        "description": "Tag that can be associated with a meal."
      },
      "PossibleMealTagCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "group": {
            "type": "string",
            "title": "Group"
          }
        },
        "type": "object",
        "required": [
          "name",
          "group"
        ],
        "title": "PossibleMealTagCreate",
        "description": "Schema for creating a possible meal tag."
      },
      "PossibleMealTagUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "group": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group"
          }
        },
        "type": "object",
        "title": "PossibleMealTagUpdate",
        "description": "Schema for updating a possible meal tag."
      },
      "TagRef": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "TagRef",
        "description": "Reference to an existing tag by ID."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}